code cleanout - static cast OR select appropriate data type
    this removes associated warnings raised during build process
    related to possible data loss which may occur
    during implicit variable conversion
    Charts\Aerolab.cpp(135): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
    Charts\AllPlot.cpp(6879): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
    Charts\CPPlot.cpp(606): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    Charts\CPPlot.cpp(607): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    Charts\CPPlot.cpp(612): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    Charts\CPPlot.cpp(2262): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    Charts\CPPlot.cpp(2263): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    Charts\CpPlotCurve.cpp(171): warning C4267: '=': conversion from 'size_t' to 'int', possible loss of data
    FileIO\MoxyDevice.cpp(548): warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data
    FileIO\SrmRideFile.cpp(163): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    FileIO\Serial.cpp(266): warning C4267: 'argument': conversion from 'size_t' to 'DWORD', possible loss of data
    FileIO\Serial.cpp(306): warning C4267: 'argument': conversion from 'size_t' to 'DWORD', possible loss of data
    FileIO\TTSReader.cpp(109): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
    FileIO\TTSReader.cpp(394): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data

+++ src/Charts/AllPlot.cpp
@@ -6876,7 +6876,7 @@ int IntervalPlotData::intervalCount() const
 double IntervalPlotData::x(size_t i) const
 {
     // for each interval there are four points, which interval is this for?
-    int interval = i ? i/4 : 0;
+    int interval = i ? static_cast<int>(i)/4 : 0;
     interval += 1; // interval numbers start at 1 not ZERO in the utility functions
 
     double multiplier = GlobalContext::context()->useMetricUnits ? 1 : MILES_PER_KM;
